backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.12
python-multipart==0.0.6
python-dotenv==1.0.0

# AI and ML libraries  
google-generativeai==0.3.2
groq==0.4.1
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.13

# Vector databases
pinecone-client==5.0.1
faiss-cpu==1.7.4

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.9.0

# HTTP and utilities
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0

# Additional utilities
numpy>=1.26.0
torch>=2.0.0
transformers>=4.35.0